Another Pyramid Test

+++ The Cherno's Tutorials/VertexArray.cpp
@@ -14,7 +14,7 @@ void vertex_Array::add_buffer(const vertex_Buffer & vb, const vertex_Buffer_Layo
     vb.bind();
     const auto & elements = vb_Layout.get_elements();
 
-    uint64_t offset = 0;
+    unsigned int offset = 0;
 
     for(unsigned int i = 0; i < elements.size(); i++){
         GLCall(glEnableVertexAttribArray(i));
